New visitor architecture, filtering and automatic downloading
Add LernplattformScraper.Visitor; LernplattformScraper's interactors now support
visiting with any class that implements that informal visitor interface.

Filtering: the user may specific a scraper config (yaml file), which will be
parsed as a filter dictionary. With it, mebis-scraper will only invoke whatever
action is specified on the subset of activities in subjects, courses, ... that
the user cares about.

-c now specifies the scraper configuration instead of the credentials file; now
-l does that instead.