Refactor: modularize acceptors and visitors, headless firefox
Move the Lernplattform*Acceptor classes, which were previously part of the main
program file (LernplattformScraper.py) to a new module of mebis_scraper
(acceptors.py). Similarly, move LernplattformFilterVisitor
mebis_scraper/filters.py.

selenium_auth is now selenium_scraping. With that change, add a new submodule to
it: profiles.py. It is is to contain various functions to create selenium
webdrivers, and make_firefox_profile was moved to it.

Extend the make_firefox_profile function to create headless selenium drivers.